pyproject.toml
# ===============================================================
#          pyproject.toml for the B-M-I Strategy
# ===============================================================
[project]
name = "permission_test"
version = "1.0.0"
# We define the dependencies as WE want them to be,
# and let the workflow handle the "hacking" to make them work.
dependencies = ["flet==0.28.3", "flet-permission-handler", "pyjnius"]

# ===============================================================
# Flet-specific build settings
# ===============================================================
[tool.flet]
product_name = "Permission Test"
[tool.flet.android]
dependencies = [
    "pyjnius"
]

[tool.flet.android.permission]
"android.permission.READ_CONTACTS" = true
"android.permission.WRITE_CONTACTS" = true

# ===============================================================
# The Critical Flutter Fix
# ===============================================================
[tool.flet.flutter.pubspec.dependency_overrides]
webview_flutter_android = "4.10.1"
